refactor: reword claim to mint
