+++ static/scripts/rewards/gift-cards/mint/mint-action.ts
